+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/r2dbc/data/UserData.java
@@ -5,6 +5,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -26,5 +27,5 @@ public class UserData {
     @Column("correo_electronico")
     private String correoElectronico;
     @Column("salario_base")
-    private Long salarioBase;
+    private BigDecimal salarioBase;
 }
